requirements.txt
numpy==1.26.4
qdrant-client==1.11.3
torch 
transformers 
streamlit  
gtts
autogen
pypdf
python-dotenv
scikit-learn
langchain
langchain-community  
huggingface-hub 
sentence-transformers
flaml[automl] 
